refactor: reposition of coordinator dependencies
