+++ src/algorithms/one_r/tester.py
@@ -1,4 +1,4 @@
-from src.tester.tester import Tester
+from src.coordinators.trainable_algorithms.tester.interface import Tester
 
 
 class OneRTester(Tester):
